requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
aiofiles==23.2.1
pillow==10.2.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
websockets==12.0
email-validator==2.1.0
agora-token-builder